fix tick icicle mixin
