+++ src/main/resources/frostiful.mixins.json
@@ -27,12 +27,12 @@
         "entity.root.EntityMovementBlockMixin",
         "entity.root.PlayerMovementBlockMixin",
         "entity.root.RootedEntityImplMixin",
-        "food.ConsumeItemCriterionMixin",
+        "item.ConsumeItemCriterionMixin",
         "item.ArmorTrimPatternsMixin",
         "item.EnchantmentEntityEffectMixin",
         "powder_snow_effects.EntityPowderSnowRedirect",
         "powder_snow_effects.LivingEntityPowderSnowDisabler",
-        "server.IcicleGrowthMixin",
+        "world.IcicleGrowthMixin",
         "world.ServerChunkManagerMixin",
         "world.WindSpawningMixin",
         "world.spawner.PatrolSpawnerMixin"
